Move object to shared dir
